learn/seminar/requirements.txt
pip==22.0.4
streamlit==1.9.0
pillow==8.4.0
easyocr==1.5.0 
altair==4.2.0 

pandas==1.3.5 # 1.4.1 
matplotlib==3.4.3 # ok
numpy==1.19.5 # 1.20
seaborn==0.11.2 # ok

base58==2.1.1
click==8
protobuf~=3.19.0 # без этого не работает из-за обновления protobuf в гугл

torch==1.10.1
torchvision==0.11.2

